Gender enum has been added into Common Enums. Inheritance of IdentityDbContext has been changed due to all addition of inherited custom Identity POCO models,  the names of these tables has been customized in TamsDbContext.cs. A new migration has been applied. IRepository interface has been added into Tams.DataAccess. EntityBase has been removed from Base folder and added into Abstracts in TAMS.Entity. AddressBase and IEntity has been removed from TAMS.Entity. AddressBase entity inheritance POCO has been removed and the EntityBase class and IAddress interface have been added into AcademyAddress, ClubAddress, and TennisTraineeAddress. Gender reference has been added into AppUser POCO. Role, RoleClaim, UserClaim, UserLogin, UserRole, and UserToken POCO classes which are inherited from the POCO classes of Identity have been added. The type of the foreign key properties have been updated from long to int in all POCO classes. IAddress, IEntity, IGeneralInfo and IKey have been added. ConnecitonStrings parameter has been deleted from appsetting.json in TAMS.Web. DbContext dependency injection has been removed on Startup.csw in TAMS.Web. TAMS.WebApi have been created as a service project.

+++ TAMS.Entity/Concrete/Club.cs
@@ -1,14 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using TAMS.Entity.Base;
+using TAMS.Entity.Abstracts;
 
 namespace TAMS.Entity.Concrete
 {
     public class Club : EntityBase
     {
         [ForeignKey("Academy")]
-        public long? AcademyId { get; set; }
+        public int? AcademyId { get; set; }
 
         [Required]
         public string ClubName { get; set; }
